+++ src/compas/files/__init__.py
@@ -68,7 +68,6 @@ from __future__ import print_function
 from .amf import *  # noqa: F401 F403
 from .dxf import *  # noqa: F401 F403
 from .gltf import *  # noqa: F401 F403
-from .gltf2 import *  # noqa: F401 F403
 from .las import *  # noqa: F401 F403
 from .obj import *  # noqa: F401 F403
 from .off import *  # noqa: F401 F403
