+++ tests/compas/files/test_gltf.py
@@ -18,3 +18,11 @@ def gltf():
 def test_from_gltf_edges_loaded(gltf):
     loaded_gltf = GLTF(gltf)
     assert len(loaded_gltf.parser.scene_data['edges']) > 0
+
+
+# gltf and bin
+#   simple
+#   with sparse
+#   with morph targets
+# gltf with data uri
+# glb
